requirements.txt
requests
websocket-client
PyYAML
loguru
pytest
typer==0.9
